+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Pods can be notified to a specific channel with `kube-slack/slack-channel` annotation on pod.
 
 ## [3.2.3] - 2018-03-10
 ### Changed